media/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
atomicwrites==1.4.0
brotlipy==0.7.0
certifi==2022.9.24
django-bootstrap4==0.0.6
django-cors-headers==3.13.0
django-filter==1.1.0
django-pure-pagination==0.3.0
djangorestframework==3.14.0
inflection==0.5.1
Markdown==3.4.1
mistune==0.8.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
Pillow==9.3.0
ply==3.11
pyls-spyder==0.4.0
PyMySQL==1.0.2
PyQt5==5.15.7
PyQtWebEngine==5.15.4
python-lsp-jsonrpc==1.0.0
pytz==2018.4
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
webencodings==0.5.1
wincertstore==0.2